Fix fitters so that the x axis can be a vector of bin centers or bin edges.
These two cases assume that the pulseheights argument has the same length as the data, or one extra value, and that the length of pulseheights can distinguish the cases. If neither length holds, then the bin numbers are used as the x axis.

This change was made to all Gaussian and fluorescence line fitters that I could find.