Mass: add a new flavor of optimal filtering with 3-lags and sometimes a sample-shift.
Uses the MicrocalDataSet._3lag_filter (which is False by default, at least for now).
So far, the 3-lag filter is the same filter as before, but you use the parabola through
the convolution only to pick p_filt_phase. The p_filt_value is chosen as the lag=0 
or lag=-1 value of the convolution, depending on whether the kink detector finds that
this pulse should have triggered one sample sooner than it did.