Change EnergyCalibration behavior for 1-2 point cal curves.
First, approximation is now allowed for 3+ points (formerly was 4+ points).
For 0-2 points, exact curves are used regardless of the approximate setting.

Second, all 2 point cal curves are treated just like 3+-point curves. They
will be given a natural spline in the space corresponding to the chosen 
curve type (though with 2 points, a natural spline is better known as a line).

Finally, for a 1-point cal, the power-law curve is only drawn for curve type 
"loglog". All other curve types assume constant gain, so they are an 
E-is-proportional-to-PH curve going through the one data point.

Fixes #63. Fixes #66.