Remove mass.mathstat.fitting.MaximumLikelihoodGaussianFitter.
This was used only in the test cases (which I've updated) and in Young's
secret, private calibration object. Too bad for him! Galen will update
its usage (see how test_fitting.py was changed for guidance). Fixes #65.

Galen: it's on you to fix so that the auto-calibration object still works.