Fix failing test due to minor change to autocorrelation calculation
Commit 5604aa0 made a small change to the logic used to decide whether a noise
record had a pulse. The logic rejects a "chunk" if it has a value too far away
from a mean value. The commit changed the calculation of the mean value to occur
at the "chunk" level instead of the "segment" level. This works better for umux
data where we have flux jumps (so that a jump causes only a single "chunk" to be
thrown away, instead of an entire segment).

Because the mean is subtracted from each chunk prior to calculating the
autocorrelation, this has the effect of making very small changes in the
autocorrelation function (because each "chunk" in a segment has a different mean
subtracted, instead of all "chunks" in a segment having the same mean
subtracted). This then causes very small changes in the filter, which leads to
very small differences in filtered pulse height (1 part in 10^6).

Since the changes in filtered value are so small, I'm going to change the
expected values to be the same as what the new code is outputting.