Add ability to truncate LJH files
This new function, ljh_truncate(), truncates an LJH file to either the first N
pulses (where N is specified by the user), or to all pulses prior to a specified
timestamp. It will copy the file header exactly, and copies the correct
timestamps and rowcounts as well as the individual traces. It does not support
LJH versions below 2.2.0.

I also added script to make it easier to call this function on a set of files.