add del data to try to prevent test failures
We're seeing test failures related to hdf5 files still being open so we can't overwrite them. This is an attempt to fix it by explicitly closing our data objects that have references to hdf5 files, so that they close the hdf5 files. In the projectors script case there is no easy handle to the data object, so I call gc.collect instead.